% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_bulk_density}
\alias{plot_bulk_density}
\title{Plot SLGA bulk density by depth}
\usage{
plot_bulk_density(bulk_density)
}
\arguments{
\item{bulk_density}{Standard bulk-density band table.}
}
\value{
The input data, invisibly.
}
\description{
Plot SLGA bulk density by depth
}
