% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_mass_balance}
\alias{plot_mass_balance}
\title{Plot numerical mass balance through time}
\usage{
plot_mass_balance(simulation)
}
\arguments{
\item{simulation}{Simulation table returned by \code{\link[=simulate_rclt]{simulate_rclt()}}.}
}
\value{
The input data, invisibly.
}
\description{
Plot numerical mass balance through time
}
