% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_observed_fitted}
\alias{plot_observed_fitted}
\title{Plot replicate observations, geometric means, and fitted concentrations}
\usage{
plot_observed_fitted(predictions)
}
\arguments{
\item{predictions}{Fit prediction table returned by \code{\link[=fit_rclt]{fit_rclt()}}.}
}
\value{
The input data, invisibly.
}
\description{
Plot replicate observations, geometric means, and fitted concentrations
}
